requirements/base.txt
fastapi[standard]==0.115.12
pydantic_settings==2.9.1
dependency-injector==4.46.0